+++ src/winxx.cc
@@ -221,9 +221,9 @@ extern "C" {
       tabs.push_back(Tab());
       Tab& tab = tabs.back();
       tab.terminal->child = tab.chld.get();
-      (term_reset)(tab.terminal.get(), true);
       tab.terminal.get()->show_scrollbar = !!cfg.scrollbar;  // hotfix #597
       (term_resize)(tab.terminal.get(), rows, cols);
+      (term_reset)(tab.terminal.get(), true);
       tab.chld->cmd = g_cmd;
       tab.chld->home = g_home;
       struct winsize wsz{rows, cols, width, height};