+++ wiki/Changelog.md
@@ -1,8 +1,11 @@
 Terminal features
   * Fix file link detection (#1208), tweak URL detection (#1209).
+  * Fix setup of initial status line if monitor DPI is not 96.
 
 Configuration
   * New user-definable function compose (#1211).
+  * New option setting ComposeKey=capslock (#1211).
+  * New option ConfirmReset to guard confirm dialog on interactive reset (#1173).
 
 ### 3.6.4 (25 Mar 2023) ###